chore: use extract environment::settings module to file
